skill-platform/python-ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyMuPDF==1.23.8
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
